Fix compile warnings

+++ src/Libpfs/vex/dotproduct.hxx
@@ -34,7 +34,7 @@ _Type dotProduct(const _Type* v1, const _Type* v2, size_t N)
 {
     double dotProd = _Type();
 #pragma omp parallel for reduction(+:dotProd)
-    for (int idx = 0; idx < N; idx++)
+    for (int idx = 0; idx < static_cast<int>(N); idx++)
     {
         dotProd += (v1[idx] * v2[idx]);
     }
@@ -46,7 +46,7 @@ _Type dotProduct(const _Type* v1, size_t N)
 {
     double dotProd = _Type();
 #pragma omp parallel for reduction(+:dotProd)
-    for (int idx = 0; idx < N; idx++)
+    for (int idx = 0; idx < static_cast<int>(N); idx++)
     {
         dotProd += (v1[idx] * v1[idx]);
     }
